+++ src/FeedKitchen.MenuWaiter/Extensions/MenuExtensions.cs
@@ -12,36 +12,30 @@ namespace FeedKitchen.MenuWaiter.Extensions
         public static SyndicationFeed Serve(this Menu menu)
         {
             var feed = new SyndicationFeed(
-                   menu.Title,
-                   menu.Description,
-                   menu.Url,
-                   menu.Url.ToString(),
+                   menu.Title, menu.Description,
+                   menu.Url, menu.Url.ToString(),
                    new DateTimeOffset(menu.LastUpdate ?? DateTime.Now));
 
             feed.Items = menu.Ingredients.Select(x =>
             {
                 var kitschenUri = !string.IsNullOrWhiteSpace(x.KitchenUri) ? new Uri(x.KitchenUri) : new Uri(x.Link);
                 var pubDate = new DateTimeOffset(x.PublishingDate ?? DateTime.Now);
-                var item = new SyndicationItem(
-                    x.Title,
-                    x.Content,
-                    kitschenUri,
-                    x.Link,
-                    pubDate)
+
+                var syndicationItem = new SyndicationItem(x.Title, x.Content, kitschenUri, x.Link, pubDate)
                 {
                     Id = x.Id,
                     Summary = new TextSyndicationContent(x.Summary, TextSyndicationContentKind.Plaintext),
                     PublishDate = pubDate
                 };
 
-                x.Categories.ForEach(y => item.Categories.Add(new SyndicationCategory(y)));
+                x.Categories.ForEach(y => syndicationItem.Categories.Add(new SyndicationCategory(y)));
 
-                item.Authors.Add(new SyndicationPerson
+                syndicationItem.Authors.Add(new SyndicationPerson
                 {
                     Name = x.Author
                 });
 
-                return item;
+                return syndicationItem;
             });
 
             return feed;
